+++ cmd/insprd/main.go
@@ -48,9 +48,10 @@ func main() {
 		AutoOffsetReset:  "earliest",
 		KafkaInsprAddr:   "http://localhost",
 		SidecarImage:     "gcr.io/red-inspr/inspr/sidecar/kafka:latest",
+		KafkaInsprPort:   3069,
 	}
 
-	brokerManager.Create(metabrokers.BrokerStatus(brokers.Kafka), config)
+	brokerManager.Create(metabrokers.BrokerStatus(metabrokers.Kafka), config)
 
 	api.Run(memoryManager, operator, authenticator, brokerManager)
 }
