+++ cmd/insprd/memory/brokers/constants.go
@@ -1,12 +1,15 @@
 package brokers
 
-import "github.com/inspr/inspr/pkg/sidecars/models"
+import (
+	"github.com/inspr/inspr/pkg/meta/brokers"
+	"github.com/inspr/inspr/pkg/sidecars/models"
+)
 
 // GetSidecarConnectionVars returns port enviroment variable
 // names for each possible broker
 func GetSidecarConnectionVars(broker string) *models.ConnectionVariables {
 	switch broker {
-	case Kafka:
+	case brokers.Kafka:
 		return &models.ConnectionVariables{
 			ReadEnvVar:  "INSPR_LBSIDECAR_READ_PORT",
 			WriteEnvVar: "INSPR_SIDECAR_KAFKA_WRITE_PORT",
